refactor: Support/xcode9 (#4)
* Support WARC 1.0 compliant WARC-Target-URI fields

Due to an error in the spec, most implementations of WARC 1.0
don't surround the contents of the WARC-Target-URI field with
angle brackets even though they're required to do so. There are
some tools that do respect this requirement, such as newer
versions of GNU Wget. Since both variants are in use, it's best
to just strip angle brackets when they're present.

Reference: https://github.com/iipc/warc-specifications/issues/23

* Support extracting warc.{gz,bz2,lzma,xz,Z} files directly

* Add travis for running tests

Co-authored-by: flan <flan@flande.re>
Co-authored-by: Paul Taykalo <ptaykalo@macpaw.com>